+++ vscode/src/configuration.ts
@@ -128,10 +128,6 @@ export function getConfiguration(
             'autocomplete.experimental.fireworksOptions',
             undefined
         ),
-        autocompleteExperimentalPreloadDebounceInterval: getHiddenSetting(
-            'autocomplete.experimental.preloadDebounceInterval',
-            0
-        ),
 
         // Note: In spirit, we try to minimize agent-specific code paths in the VSC extension.
         // We currently use this flag for the agent to provide more helpful error messages